% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_password.R
\name{generate_password}
\alias{generate_password}
\title{Generate a random password}
\usage{
generate_password(
  length,
  uppercase = TRUE,
  lowercase = TRUE,
  numbers = TRUE,
  special_chars = TRUE
)
}
\arguments{
\item{length}{The desired length of the password}

\item{uppercase}{Boolean indicating whether to include uppercase letters}

\item{lowercase}{Boolean indicating whether to include lowercase letters}

\item{numbers}{Boolean indicating whether to include numbers}

\item{special_chars}{Boolean indicating whether to include special characters}
}
\value{
A randomly generated password
}
\description{
Generate a random password
}
\examples{
generate_password(10, uppercase = TRUE, lowercase = FALSE, numbers = TRUE, special_chars = TRUE)

}
