% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palindrome.R
\name{is_palindrome}
\alias{is_palindrome}
\title{Check if a string is a palindrome}
\usage{
is_palindrome(text)
}
\arguments{
\item{text}{The input string}
}
\value{
TRUE if the string is a palindrome, FALSE otherwise
}
\description{
This function checks if a given string is a palindrome (reads the same forwards and backwards).
}
