% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reverse_String.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse a string}
\usage{
reverse_string(text)
}
\arguments{
\item{text}{The input string}
}
\value{
The reversed string
}
\description{
Reverse a string
}
