% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiffle_string.R
\name{shuffle_string}
\alias{shuffle_string}
\title{shuffle_string

Shuffle characters in a string}
\usage{
shuffle_string(text)
}
\arguments{
\item{text}{The input string "any text you want to put"}
}
\value{
The shuffled string
}
\description{
shuffle_string

Shuffle characters in a string
}
