% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCount.R
\name{getCount}
\alias{getCount}
\title{getCount() function that counts keywords}
\usage{
getCount(data, keyword)
}
\arguments{
\item{data}{Input any dataset you want}

\item{keyword}{Any word or string you want to count}
}
\value{
A numeric value for the count of the specified keyword or words
}
\description{
getCount() function that counts keywords
}
\examples{
getCount(data, 'test')


}
